Fixed the incorrect type casting of the elements.
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@8553 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ src/com/l7tech/console/table/MimePartsTableSorter.java
@@ -272,7 +272,7 @@ public class MimePartsTableSorter  extends FilteredDefaultTableModel {
                     }
                 } else {
                      if (elementA instanceof Integer) {
-                        return ((Integer) elementB).intValue() > ((Long) elementA).intValue()?1:0;
+                        return ((Integer) elementB).intValue() > ((Integer) elementA).intValue()?1:0;
                     } else if(elementA instanceof String) {
                         return ((String)elementB).compareToIgnoreCase((String)elementA);
                     } else {
